requirements.txt
kagglehub
pandas
matplotlib
opencv-python-headless
tqdm
tensorflow
numpy
glob2
streamlit
pillow